Expose `AbstractDependencyFilter` from `internal` to `public`
